refactor: move modern detection from server to utils
