+++ packages/utils/package.json
@@ -12,8 +12,10 @@
     "fs-extra": "^7.0.1",
     "hash-sum": "^1.0.2",
     "proper-lockfile": "^4.1.1",
+    "semver": "^6.0.0",
     "serialize-javascript": "^1.7.0",
-    "signal-exit": "^3.0.2"
+    "signal-exit": "^3.0.2",
+    "ua-parser-js": "^0.7.19"
   },
   "publishConfig": {
     "access": "public"
