+++ packages/utils/src/index.js
@@ -7,3 +7,4 @@ export * from './serialize'
 export * from './task'
 export * from './timer'
 export * from './cjs'
+export * from './modern'
